\name{competition_codes}
\alias{competition_codes}
\title{Returns dataframe of available Competition codes and seasons available in the load functions}
\usage{
competition_codes()
}
\description{
Returns dataframe of available Competition codes and seasons available in the load_player_competition_stats, load_player_game_stats, and load_bbb_data functions.
}
\examples{
competition_codes()
}
